rename EnrichGdprDialog to EnrichGdprSidebar
