+++ frontend/src/components/Mining/EnrichGdprSidebar.vue
